refactor(fsl-ui)!: rename the chips to the reference split; close F-041 from the docs
BREAKING CHANGE: `Badge` is now the descriptive chip (Structure,
informational colours) and the valence status pill is `StatusLight`
(Feedback). `Chip` is removed — it folds into `Badge`. The published
scopes move with the names: `chip` → `badge`, `badge` → `status-light`.

F-040: both reference systems give the descriptive chip the word
`Badge` and the status member another name — Spectrum's StatusLight,
Chakra's Status. Ours were inverted, which in an AI-first package is a
first-pass-correctness cost: an agent asked for "a role chip" reaches
for `Badge` on priors from every other library and got a valence it
never asked for. The owner took the rename now rather than at the
version boundary, and now was the cheaper half of the trade the entry
framed: `Chip` had existed for hours with three consumers, so this was
two source files, two suites, two stories and six Studio sites, all
mechanical. Waiting adds every consumer written in between.

`CHIP_BOX` keeps its name deliberately. It is no longer any component's,
which is right for a shared source that names the physical object two
components render while the entity decides only its colours.

F-041 closed without building anything, because `breakpoints.md` had
already placed it and the entry was filed without reading it. The family
doc calls breakpoints adaptation infrastructure with no semantic layer,
says applications may adjust or replace them, and names this exact case:
local aliases such as `navCollapse` stay in the application layer. The
same page states the rule the shell's API already follows — breakpoints
define when layout changes, not how components behave — so neither
package should ship the hook.

F-042 is what the attempt found. Wiring Meridian to the narrow shell is
the first time two recorded workarounds met: the drawer opens with its
accessible name and the tablist inside it renders empty, because React
Aria does not populate a collection portaled into a Modal. Meridian's
sidebar is a TabList only because `Link` cannot mark `aria-current`
(F-002), and the frame lives inside one `Tabs` because a panel-less
TabList emits invalid ARIA (F-017); a collection owner and a portal
cannot be the same element. The hook and the wiring were written and
then removed rather than shipped half-working — hiding the navigation on
a phone is worse than the scroll F-023 measured.

Kept from the attempt: the Studio's `matchMedia` stub answered false to
everything, which was harmless while only `prefers-color-scheme` asked
and would have silently collapsed the shell in every test the moment
anything asked about width. It now answers `min-width` against jsdom's
actual viewport.

Refs: F-040, F-041, F-042
