+++ docs/fsl-storybook/stories/structure/Badge.stories.tsx
@@ -1,14 +1,14 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
-import { Badge, Chip, Stack, Text } from '@ttoss/fsl-ui';
+import { Badge, Stack, StatusLight, Text } from '@ttoss/fsl-ui';
 
-const meta: Meta<typeof Chip> = {
-  title: 'Structure/Chip',
-  component: Chip,
+const meta: Meta<typeof Badge> = {
+  title: 'Structure/Badge',
+  component: Badge,
 };
 
 export default meta;
 
-type Story = StoryObj<typeof Chip>;
+type Story = StoryObj<typeof Badge>;
 
 export const Default: Story = {
   args: { children: 'Admin' },
@@ -18,8 +18,8 @@ export const Emphasis: Story = {
   render: () => {
     return (
       <Stack direction="horizontal" gap="sm" align="center">
-        <Chip evaluation="muted">Muted</Chip>
-        <Chip evaluation="primary">Primary</Chip>
+        <Badge evaluation="muted">Muted</Badge>
+        <Badge evaluation="primary">Primary</Badge>
       </Stack>
     );
   },
@@ -27,29 +27,29 @@ export const Emphasis: Story = {
 
 /**
  * The distinction the component exists for. Both chips are the same box — a
- * `Chip` beside a `Badge` must not disagree about its own roundness — and they
+ * `Badge` beside a `StatusLight` must not disagree about its own roundness — and they
  * say different things: the Badge reports an outcome the system observed, the
- * Chip labels content that simply is what it is.
+ * Badge labels content that simply is what it is.
  */
-export const AgainstBadge: Story = {
+export const AgainstStatusLight: Story = {
   render: () => {
     return (
       <Stack gap="md">
         <Stack direction="horizontal" gap="sm" align="center">
           <Text variant="label-sm" tone="muted">
-            Chip — descriptive, no outcome
+            Badge — descriptive, no outcome
           </Text>
-          <Chip>Admin</Chip>
-          <Chip>Editor</Chip>
-          <Chip>Beta</Chip>
+          <Badge>Admin</Badge>
+          <Badge>Editor</Badge>
+          <Badge>Beta</Badge>
         </Stack>
         <Stack direction="horizontal" gap="sm" align="center">
           <Text variant="label-sm" tone="muted">
-            Badge — an outcome the system reports
+            StatusLight — an outcome the system reports
           </Text>
-          <Badge evaluation="positive">Passing</Badge>
-          <Badge evaluation="caution">Degraded</Badge>
-          <Badge evaluation="negative">Failing</Badge>
+          <StatusLight evaluation="positive">Passing</StatusLight>
+          <StatusLight evaluation="caution">Degraded</StatusLight>
+          <StatusLight evaluation="negative">Failing</StatusLight>
         </Stack>
       </Stack>
     );
@@ -73,7 +73,7 @@ export const InAList: Story = {
               align="center"
             >
               <Text>{member.name}</Text>
-              <Chip>{member.role}</Chip>
+              <Badge>{member.role}</Badge>
             </Stack>
           );
         })}
